dghyp.R: Minor cosmetics.
dhyperb.R: Added in hyperbBreaks(). This was earlier removed due to the
fact that it is not called by any other function. However, it was put
back in as it is more consistent to use hyperb functions where possible
instead of mixing between equivalent hyperb and ghyp functions.

dghyp.Rd: Changed documentation regarding the Theta parameter, simply
showing the greek symbols instead of the names of the characters.

dhyperb.Rd: Added in documentation regarding hyperbBreaks() parameter.
Fixed example function calls. Changed references to old (pi, zeta) to
(alpha, beta).